+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # CHANGELOG
 
+## Unreleased
+
+
 ## [0.8.10](https://github.com/uhppoted/uhppote-cli/releases/tag/v0.8.10) - 2025-01-30
 
 ### Added